+++ src/Cassowary/ConstraintData.cs
@@ -1,3 +1,3 @@
 ﻿namespace Cassowary;
 
-internal record struct ConstraintData(Expression Expression, float Strength, RelationalOperator Operator);
+internal record struct ConstraintData(Expression Expression, double Strength, RelationalOperator Operator);
